+++ MoneyExchange.Application.DTO/ReturnExchangeDto.cs
@@ -2,9 +2,9 @@
 {
     public class ReturnExchangeDto
     {
-        public decimal MontoCambiado { get; set; }
+        public double MontoCambiado { get; set; }
         public string MonedaOrigen { get; set; }
         public string MonedaDestino { get; set; }
-        public decimal TipoCambio { get; set; }
+        public double TipoCambio { get; set; }
     }
 }
